+++ src/meltano/cli/params.py
@@ -1,12 +1,22 @@
 import os
 import functools
 import click
+import urllib
 
+from meltano.core.project import Project
 from meltano.core.utils import pop_all
-from meltano.core.db import DB
 
 
 def db_options(func):
+    @click.option(
+        "-B",
+        "--backend",
+        envvar="MELTANO_BACKEND",
+        default="sqlite",
+        type=click.Choice(["sqlite", "postgresql"]),
+        help="Database backend for Meltano.",
+    )
+    @click.option("--path", default="meltano.db")
     @click.option(
         "-H",
         "--host",
@@ -32,8 +42,26 @@ def db_options(func):
     @click.password_option(envvar="PG_PASSWORD")
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
-        config = pop_all(("host", "port", "database", "username", "password"), kwargs)
-        DB.setup(**config)
-        return func(*args, **kwargs)
+        engine_uri = os.getenv("SQL_ENGINE_URI")
+        backend = kwargs.pop("backend")
+
+        config = {
+            "sqlite": pop_all(("path",), kwargs),
+            "postgresql": pop_all(
+                ("host", "port", "database", "username", "password"), kwargs
+            ),
+        }
+
+        if not engine_uri and backend == "sqlite" and config[backend]["path"]:
+            engine_uri = "sqlite:///{path}".format(**config[backend])
+
+        if not engine_uri and backend == "postgresql":
+            pg_config = config[backend]
+            pg_config["password"] = urllib.parse.quote(pg_config["password"])
+            engine_uri = "postgresql://{username}:{password}@{host}:{port}/{database}".format(
+                **pg_config
+            )
+
+        return func(*args, **kwargs, engine_uri=engine_uri)
 
     return wrapper