+++ src/meltano/cli/schema.py
@@ -3,13 +3,13 @@ import click
 import psycopg2
 import psycopg2.sql
 
-from meltano.core.db import db_open
+from meltano.core.db import DB, project_engine
+from meltano.core.project import Project
 from . import cli
 from .params import db_options
 
 
 @cli.group()
-@db_options
 def schema():
     pass
 
@@ -17,6 +17,9 @@ def schema():
 @schema.command()
 @click.argument("schema")
 @click.argument("roles", nargs=-1, required=True)
-def create(schema, roles):
-    with db_open() as db:
-        db.ensure_schema_exists(schema, grant_roles=roles)
+@db_options
+def create(schema, roles, engine_uri):
+    project = Project.find()
+    engine, _ = project_engine(project, engine_uri)
+
+    DB.ensure_schema_exists(engine, schema, grant_roles=roles)