+++ src/meltano/core/db.py
@@ -3,104 +3,97 @@ import os
 import contextlib
 import sqlalchemy.pool as pool
 import logging
+import weakref
 
 from sqlalchemy import create_engine, MetaData
-from sqlalchemy.orm import sessionmaker, scoped_session
+from sqlalchemy.orm import sessionmaker
 from sqlalchemy.sql import text
 from sqlalchemy.ext.declarative import declarative_base
 from psycopg2.sql import Identifier, SQL
 
 
-def engine_uri(**db_config):
-    return "postgresql://{user}:{password}@{host}:{port}/{database}".format(**db_config)
+SystemMetadata = MetaData()
+SystemModel = declarative_base(metadata=SystemMetadata)
+
+# Keep a Project → Engine mapping to serve
+# the same engine for the same Project
+_engines = dict()
 
 
-SystemModel = declarative_base(metadata=MetaData(schema="meltano"))
+def project_engine(project, engine_uri=None, default=False):
+    """Creates and register a SQLAlchemy engine for a Meltano project instance."""
+
+    # return the default engine if it is registered
+    if not engine_uri and project in _engines:
+        return _engines[project]
+    elif (project, engine_uri) in _engines:
+        return _engines[(project, engine_uri)]
+
+    if not engine_uri:
+        raise ValueError("No engine registered for this project.")
+
+    engine = create_engine(engine_uri)
+    Session = sessionmaker(bind=engine)
+
+    if default:
+        # register the default engine
+        _engines[project] = engine, Session
+    else:
+        _engines[(project, engine_uri)] = engine, Session
+
+    init_hook(engine)
+
+    return engine, Session
 
 
-class MetaDB(type):
-    def __init__(cls, *_):
-        cls._default = None
+def init_hook(engine):
+    function_map = {"sqlite": init_sqlite_hook, "postgresql": init_postgresql_hook}
 
-    @property
-    def default(cls):
-        return cls._default
-
-    @default.setter
-    def default(cls, v):
-        if cls._default:
-            cls._default.close()
-
-        cls._default = v
+    try:
+        function_map[engine.dialect.name](engine)
+    except KeyError:
+        raise Exception("Meltano only supports SQLite and PostgreSQL")
+    except:
+        logging.fatal("Can't initialize database.")
 
 
-class DB(metaclass=MetaDB):
-    db_config = {
-        "host": os.getenv("PG_ADDRESS", "localhost"),
-        "port": os.getenv("PG_PORT", 5432),
-        "user": os.getenv("PG_USERNAME", os.getenv("USER")),
-        "password": os.getenv("PG_PASSWORD"),
-        "database": os.getenv("PG_DATABASE"),
-    }
+def init_postgresql_hook(engine):
+    schema = os.getenv("PG_SCHEMA", "meltano")
+    DB.ensure_schema_exists(engine, schema)
+
+    SystemMetadata.schema = schema
+
+    # we need to manually set the schema for Table that
+    # are already imported.
+    for table in SystemMetadata.tables.values():
+        table.schema = schema
+
+    seed(engine)
+
+
+def init_sqlite_hook(engine):
+    seed(engine)
+
+
+# TODO: alembic hook?
+def seed(engine):
+    # import all the models
+    import meltano.core.job
+
+    # seed the database
+    SystemMetadata.create_all(engine)
+
+
+class DB:
+    @classmethod
+    def create_database(cls, engine, db_name):
+        """
+        Create a new database, PostgreSQL only.
+        """
+        engine.execute(text(f"CREATE DATABASE {db_name}"))
 
     @classmethod
-    def setup(cls, **kwargs):
-        """
-        Store the DB connection parameters and create a default engine.
-        """
-        cls.db_config.update(
-            {k: kwargs[k] for k in cls.db_config.keys() if k in kwargs}
-        )
-
-        # use connection pooling for all connections
-        pool.manage(psycopg2)
-        cls.default = cls()
-
-    @classmethod
-    def create_database(cls, db_name):
-        """
-        Create a new database
-        """
-        config = cls.db_config.copy()
-        config.update({"database": "postgres"})
-        with create_engine(
-            engine_uri(**config), isolation_level="AUTOCOMMIT"
-        ).connect() as con:
-            con.execute(text(f"CREATE DATABASE {db_name}"))
-
-    @classmethod
-    def close_default(cls):
-        """
-        Close the default engine
-        """
-        if cls.default:
-            cls.default.close()
-
-    @property
-    def engine(self):
-        return self._engine
-
-    def __init__(self):
-        self._engine = create_engine(engine_uri(**self.db_config))
-        self._session_cls = scoped_session(sessionmaker(bind=self.engine))
-
-    def create_connection(self):
-        return psycopg2.connect(**self.db_config)
-
-    def create_session(self):
-        return self._session_cls()
-
-    def open(self):
-        return db_open(self)
-
-    def session(self):
-        return session_open(self)
-
-    def close(self):
-        if self.engine:
-            self.engine.dispose()
-
-    def ensure_schema_exists(self, schema_name, grant_roles=()):
+    def ensure_schema_exists(cls, engine, schema_name, grant_roles=()):
         """
         Make sure that the given schema_name exists in the database
         If not, create it
@@ -108,61 +101,23 @@ class DB(metaclass=MetaDB):
         :param db_conn: psycopg2 database connection
         :param schema_name: database schema
         """
-        conn = self.create_connection()
-        cursor = conn.cursor()
-        schema_identifier = Identifier(schema_name)
-        group_identifiers = SQL(",").join(map(Identifier, grant_roles))
+        schema_identifier = schema_name
+        group_identifiers = ",".join(grant_roles)
 
-        create_schema = SQL(
-            """
-        CREATE SCHEMA IF NOT EXISTS {}
-        """
-        ).format(schema_identifier)
+        create_schema = text(f"CREATE SCHEMA IF NOT EXISTS {schema_identifier}")
+        grant_select_schema = text(
+            f"ALTER DEFAULT PRIVILEGES IN SCHEMA {schema_identifier} GRANT SELECT ON TABLES TO {group_identifiers}"
+        )
+        grant_usage_schema = text(
+            f"GRANT USAGE ON SCHEMA {schema_identifier} TO {group_identifiers}"
+        )
 
-        grant_select_schema = SQL(
-            """
-        ALTER DEFAULT PRIVILEGES IN SCHEMA {} GRANT SELECT ON TABLES TO {}
-        """
-        ).format(schema_identifier, group_identifiers)
-
-        grant_usage_schema = SQL(
-            """
-        GRANT USAGE ON SCHEMA {} TO {}
-        """
-        ).format(schema_identifier, group_identifiers)
-
-        cursor.execute(create_schema)
-        if grant_roles:
-            cursor.execute(grant_select_schema)
-            cursor.execute(grant_usage_schema)
-
-        conn.commit()
+        with engine.connect() as conn, conn.begin():
+            conn.execute(create_schema)
+            if grant_roles:
+                conn.execute(grant_select_schema)
+                conn.execute(grant_usage_schema)
 
         logging.info("Schema {} has been created successfully.".format(schema_name))
         for role in grant_roles:
             logging.info("Usage has been granted for role: {}.".format(role))
-
-
-@contextlib.contextmanager
-def db_open(db: DB = None):
-    """Provide a raw connection in a transaction."""
-    connection = (db or DB.default).create_connection()
-
-    try:
-        yield connection
-        connection.commit()
-    except Exception:
-        connection.rollback()
-        raise
-
-
-@contextlib.contextmanager
-def session_open(db: DB = None):
-    """Provide a transactional scope around a series of operations."""
-    session = (db or DB.default).create_session()
-    try:
-        yield session
-        session.commit()
-    except Exception:
-        session.rollback()
-        raise