+++ src/meltano/core/runner/__init__.py
@@ -1,14 +1,3 @@
-from meltano.core.db import DB, SystemModel
-
-
 class Runner:
-    def before_run(self, *args, **kwargs):
-        DB.default.ensure_schema_exists("meltano")
-        SystemModel.metadata.create_all(DB.default.engine)
-
-    def run(self, *args, **kwargs):
+    def run(self):
         pass
-
-    def perform(self, *args, **kwargs):
-        self.before_run(*args, **kwargs)
-        self.run(*args, **kwargs)