+++ CHANGELOG.md
@@ -13,8 +13,10 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Changes
 
 ### Fixes
+* Meltano will correctly source the `.env`
 
 ### Breaks
+* Meltano will now use SQLite for the job log. See https://meltano.com/docs/architecture.html#job-logging for more details.
 
 
 ## 0.9.0 - (2019-02-05)